requirements.txt
aiofiles==23.1.0
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
certifi==2022.12.7
charset-normalizer==3.0.1
colorama==0.4.6
flake8==6.0.0
frozenlist==1.3.3
googletrans-py==4.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
loguru==0.7.0
magic-filter==1.0.9
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==0.4.3
peewee==3.15.4
pycodestyle==2.10.0
pydantic==1.10.10
pyflakes==3.0.1
pyTelegramBotAPI==4.9.0
python-dotenv==1.0.0
pytz==2023.3
redis==4.6.0
requests==2.31.0
rfc3986==1.5.0
schedule==1.1.0
sniffio==1.3.0
tomli==2.0.1
types-peewee==3.15.0.5
types-requests==2.28.11.8
types-urllib3==1.26.25.4
typing_extensions==4.4.0
urllib3==1.26.14
win32-setctime==1.1.0
yarl==1.9.2
